# Small optimization for std.string.capitalize