Previous code had to iterate over the array twice and allocated two strings. 

Performance:

```
$ ldc2 -O5 -release test.d && ./test

old	816 ms, 65 μs, and 6 hnsecs
new	506 ms, 437 μs, and 7 hnsecs
```

code

```d
import std.array;
import std.traits;
import std.stdio;
import std.datetime;
import std.conv;
import std.range;
import std.algorithm;

enum testCount = 1_000_000;

S capitalize1(S)(S input) @trusted pure
if (isSomeString!S)
{
    import std.array : array;
    import std.conv : to;
    import std.uni : asCapitalized;

    return input.asCapitalized.array.to!S;
}

S capitalize2(S)(S input) @trusted pure
if (isSomeString!S)
{
    import std.array : array;
    import std.utf : byUTF;
    import std.uni : asCapitalized;

    return input.asCapitalized.byUTF!(ElementEncodingType!(S)).array;
}

void main()
{
    auto result = to!Duration(benchmark!(() => capitalize1(`this is a test`))(testCount)[0]);
    auto result2 = to!Duration(benchmark!(() => capitalize2(`this is a test`))(testCount)[0]);

    writeln("old", "\t", result);
    writeln("new", "\t", result2);
}
```