@JackStouffer 2017-02-22
Auto-merge toggled on